requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
celery>=5.3.4
redis>=5.0.1
aiogram>=3.4.0
aiogram-dialog>=2.0.0
python-dotenv>=1.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.1